requirements.txt
request
faster-whisper